registration and comments

+++ authors/index.php
@@ -1,5 +1,5 @@
 <?php
-
+session_start();
 require_once '../config/connect.php';
 
 ?>
@@ -27,29 +27,33 @@ require_once '../config/connect.php';
         <a href="../queries/index.php" class="item">Filter</a>
 
     </div>
-
-    <div class="ui black segment">
-        <form action="addAuthor.php" method="post" class="ui form">
-            <div class="fields">
-                <div class="field" style="flex-grow: 1">
-                    <label>Full name</label>
-                    <input type="text" placeholder="Full name" name="full_name">
+    <?php
+    if ($_SESSION['user']['role'] === 'admin') {
+        ?>
+        <div class="ui black segment">
+            <form action="addAuthor.php" method="post" class="ui form">
+                <div class="fields">
+                    <div class="field" style="flex-grow: 1">
+                        <label>Full name</label>
+                        <input type="text" placeholder="Full name" name="full_name">
+                    </div>
+                    <div class="field">
+                        <label>Date of birth</label>
+                        <input type="date" name="date_of_birth">
+                    </div>
+                    <div class="field">
+                        <label>Date of death</label>
+                        <input type="date" name="date_of_death">
+                    </div>
+                    <button type="submit" class="ui icon button" style="align-self: flex-end; margin-bottom: 2px;">
+                        <i class="plus icon"></i>
+                    </button>
                 </div>
-                <div class="field">
-                    <label>Date of birth</label>
-                    <input type="date" name="date_of_birth">
-                </div>
-                <div class="field">
-                    <label>Date of death</label>
-                    <input type="date" name="date_of_death">
-                </div>
-                <button type="submit" class="ui icon button" style="align-self: flex-end; margin-bottom: 2px;">
-                    <i class="plus icon"></i>
-                </button>
-            </div>
-        </form>
-    </div>
-
+            </form>
+        </div>
+        <?php
+    }
+    ?>
     <table class="ui celled table">
         <thead>
         <tr>
@@ -57,8 +61,14 @@ require_once '../config/connect.php';
             <th>Full name</th>
             <th>Date of birth</th>
             <th>Date of death</th>
-            <th>Delete</th>
-            <th>Edit</th>
+            <?php
+            if ($_SESSION['user']['role'] === 'admin') {
+                ?>
+                <th>Delete</th>
+                <th>Edit</th>
+                <?php
+            }
+            ?>
         </tr>
         </thead>
         <tbody>
@@ -68,10 +78,13 @@ require_once '../config/connect.php';
         foreach ($authors as $author) {
             ?>
             <tr>
-                <td><?= $author[0] ?></td>
-                <td><?= $author[1] ?></td>
-                <td><?= $author[2] ?></td>
-                <td><?= $author[3] ?></td>
+            <td><?= $author[0] ?></td>
+            <td><?= $author[1] ?></td>
+            <td><?= $author[2] ?></td>
+            <td><?= $author[3] ?></td>
+            <?php
+            if ($_SESSION['user']['role'] === 'admin') {
+                ?>
                 <td style="text-align: center">
                     <a href="deleteAuthor.php?id=<?= $author[0] ?>">
                         <div class="ui vertical animated button" tabindex="0">
@@ -92,8 +105,9 @@ require_once '../config/connect.php';
                         </div>
                     </a>
                 </td>
-            </tr>
-            <?php
+                </tr>
+                <?php
+            }
         }
         ?>
         </tbody>
